Code_updated

+++ src/margi/CodingTest/CodingTest_5.java
@@ -4,7 +4,7 @@ public class CodingTest_5 {
 
 	void getPerfectNumbers() {
 		int count=0;
-		long sum, number = 1;
+		int sum, number = 1;
 		while(count<5) {
 			sum=0;
 			for(int index=1; index<number; index++) {
